backend/requirements.txt
fastapi
uvicorn
black
python-dotenv
requests
matplotlib
google-genai
